+++ src/main/java/com/jorgetfg/backend/controllers/SubscriptionController.java
@@ -2,12 +2,12 @@ package com.jorgetfg.backend.controllers;
 
 import com.jorgetfg.backend.dto.CompleteSubscriptionDto;
 import com.jorgetfg.backend.dto.CreateSubscriptionDto;
-import com.jorgetfg.backend.services.SubscriptionService;
+import com.jorgetfg.backend.services.Imp.SubscriptionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Set;
+import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
@@ -16,13 +16,21 @@ public class SubscriptionController {
     private final SubscriptionService subscriptionService;
 
     @GetMapping(path = "/users/{userId}/subscriptions")
-    public ResponseEntity<Set<CompleteSubscriptionDto>> getUserSubscriptions(@PathVariable Long userId) {
+    public ResponseEntity<List<CompleteSubscriptionDto>> getUserSubscriptions(@PathVariable Long userId) {
+        List<CompleteSubscriptionDto> completeSubscriptionDtoList = subscriptionService.getUserSubscriptions(userId);
+        if (completeSubscriptionDtoList == null) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(subscriptionService.getUserSubscriptions(userId));
     }
 
     @GetMapping(path = "/users/{userId}/subscriptions/{subscriptionId}")
     public ResponseEntity<CompleteSubscriptionDto> getUserSubscription(@PathVariable Long userId,
                                                                        @PathVariable Long subscriptionId) {
+        CompleteSubscriptionDto completeSubscriptionDto =  subscriptionService.getUserSubscription(userId,subscriptionId);
+        if (completeSubscriptionDto == null) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(subscriptionService.getUserSubscription(userId,subscriptionId));
     }
 
@@ -30,6 +38,11 @@ public class SubscriptionController {
     public ResponseEntity<CompleteSubscriptionDto> addUserSubscription(@PathVariable Long userId,
                                                                        @PathVariable Long folderId,
                                                                        @RequestBody CreateSubscriptionDto createSubscriptionDto) {
+        CompleteSubscriptionDto completeSubscriptionDto = subscriptionService.addUserSubscription(userId,folderId,
+                createSubscriptionDto);
+        if (completeSubscriptionDto == null) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(subscriptionService.addUserSubscription(userId,folderId,createSubscriptionDto));
     }
 
@@ -38,6 +51,11 @@ public class SubscriptionController {
                                                                           @PathVariable Long folderId,
                                                                           @PathVariable Long subscriptionId,
                                                                           @RequestBody CreateSubscriptionDto createSubscriptionDto) {
+        CompleteSubscriptionDto completeSubscriptionDto = subscriptionService.updateUserSubscription(userId,folderId,
+                subscriptionId,createSubscriptionDto);
+        if (completeSubscriptionDto == null) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(subscriptionService.updateUserSubscription(userId,folderId,subscriptionId,createSubscriptionDto));
     }
 
@@ -45,6 +63,10 @@ public class SubscriptionController {
     public ResponseEntity<CompleteSubscriptionDto> deleteUserSubscription(@PathVariable Long userId,
                                                                           @PathVariable Long folderId,
                                                                           @PathVariable Long subscriptionId) {
+        CompleteSubscriptionDto completeSubscriptionDto = subscriptionService.deleteUserSubscription(userId,folderId,subscriptionId);
+        if (completeSubscriptionDto == null) {
+            return ResponseEntity.badRequest().build();
+        }
         return ResponseEntity.ok(subscriptionService.deleteUserSubscription(userId,folderId,subscriptionId));
     }
 }
